Change deletion of files and directories from a 2-pass process to a single-pass process
Previous code would try recursively deleting the directory tree first (only the flagged directories), and if any files/non-flagged directories existed it would fail; the failure would then lead to processing the directory to handle flagged-file deletions (only). The next time the update manager ran, it would then be able to delete the directory tree, if it was now empty.

The updated code processes the subdirectory first to look for files that need to be deleted, and tries deleting flagged directories after all nested file/directory deletions have occurred

This does not handle 'root' directories that have been flagged for deletion. Current code ignores 'root' directories that are flagged for deletion and still contain files/directories, even if all of them are flagged for deletion.