Perform a possible two-pass recursive deletion for a flagged 'root' directory
Existing code could only handle recursive empty-directory (flagged) deletion, and never did any processing to handle deleting flagged nested files. New code, if the 'root' directory is flagged for deletion and contains files/non-flagged folders, will then go on to process the contents of that directory to handle subdirectories/files that are flagged for deletion. Code will not delete anything that is or contains a non-flagged file/directory.